% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{alignment_check}
\alias{alignment_check}
\title{check the alignment of GWAS to reference panel and flip allele beta where required}
\usage{
alignment_check(df, reference, bfile)
}
\arguments{
\item{df}{A data frame containing the GWAS summary statistics.
The following columns should be present:
\itemize{
\item \strong{CHR}: Chromosome number (integer).
\item \strong{POS}: Position of the SNP (integer).
\item \strong{SNP}: SNP identifier (character).
\item \strong{EA}: Effect allele (character).
\item \strong{OA}: Other allele (character).
\item \strong{EAF}: Effect allele frequency (numeric).
\item \strong{BETA}: Effect size estimate (numeric).
\item \strong{SE}: Standard error of the effect size (numeric).
\item \strong{P}: P-value for the association (numeric).
\item \strong{N}: Sample size (integer).
\item \strong{phenotype}: Phenotype identifier (character).
}}

\item{reference}{A data frame containing the reference data for comparison.
The following columns should be present:
\itemize{
\item \strong{Predictor}: SNP identifier (character).
\item \strong{A1}: Allele 1 (character).
\item \strong{A2}: Allele 2 (character).
\item \strong{A1_Mean}: Mean value for allele 1 (numeric).
\item \strong{MAF}: Minor allele frequency (numeric).
\item \strong{Call_Rate}: Call rate (integer).
\item \strong{Info}: Information score (integer).
}}

\item{bfile}{file path for reference population (built for using 1kG,
e.g., /path/EUR/EUR).}
}
\value{
A list containing:
\itemize{
\item \strong{plots}: A list of plots, including the alignment plot and
observed vs expected z-score plots.
\item \strong{list_df}: The final data frame after allele flipping and
adjustments.
\item \strong{lambda}: A list of lambda estimates for the adjusted z-scores.
}
}
\description{
This function performs an alignment check for GWAS data by comparing the input
data (\code{df}) with a reference dataset (\code{reference}). It checks if the alleles
in the data are aligned and flips the alleles when necessary to ensure
consistency with the reference. The function also computes an LD matrix,
performs a kriging-based procedure to adjust the z-scores, and generates a
series of plots to visualize the alignment.
}
\details{
The function first merges the GWAS summary statistics with the reference data
based on the SNP identifier and flips the effect sizes if the effect allele
(EA) does not match allele 1 (A1) in the reference. It then computes an LD
matrix using the \code{ieugwasr::ld_matrix_local} function, which is used in
subsequent analyses. The function also runs a kriging procedure using
\code{susieR::kriging_rss} to adjust the z-scores based on the LD matrix, and
generates plots comparing the observed and expected z-scores before and
after allele flipping. The function will iteratively flip alleles and update
the data until no moreallele flips are needed (based on a log likelihood
ratio test).
}
