% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{biomaRt_getBM_batch}
\alias{biomaRt_getBM_batch}
\title{Retrieve Batched Data from BioMart with Filters and Attribute Chunks}
\usage{
biomaRt_getBM_batch(mart, attributes, filters, values, chunk_size)
}
\arguments{
\item{mart}{A \code{Mart} object created with \code{useMart}, representing the BioMart dataset connection.}

\item{attributes}{A character vector of attributes to retrieve from BioMart.}

\item{filters}{A single attribute (string) used as a filter in the BioMart query, often the unique identifier.}

\item{values}{A vector of values to match with the specified \code{filters} attribute.}

\item{chunk_size}{An integer defining the number of attributes per batch request.}
}
\value{
A data frame containing the retrieved and filtered data, with each row representing a unique entity
and columns corresponding to the requested attributes.
}
\description{
This function retrieves data from BioMart in batches by splitting attributes into smaller chunks.
It fetches data using specified filters and joins the results sequentially, with optional filtering for protein-coding genes
and standard chromosomes.
}
\details{
This function processes the retrieval in chunks, making sequential calls to BioMart and avoiding large single queries.
The resulting data is filtered to include only protein-coding genes and standard chromosomes (1–22, X, Y).
If an attribute specified as \code{filters} is also in \code{attributes}, it will be removed from \code{attributes} to prevent duplication.
}
\examples{
\dontrun{
library(biomaRt)
mart <- useMart("ensembl", dataset = "hsapiens_gene_ensembl")
attributes <- c("chromosome_name", "start_position", "end_position", "hgnc_symbol")
filters <- "hgnc_symbol"
values <- c("BRCA1", "TP53", "EGFR")
chunk_size <- 2
result <- biomaRt_getBM_batch(mart, attributes, filters, values, chunk_size)
}

}
