% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_plot.R
\name{blank_plot}
\alias{blank_plot}
\title{Create a blank ggplot2 plot}
\usage{
blank_plot(message)
}
\arguments{
\item{message}{a message that no plot was made}
}
\value{
A ggplot2 plot object
}
\description{
This function creates a blank ggplot2 plot with no data or aesthetics
}
