% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-change-insertion.R
\name{calculate_and_insert_percent_change}
\alias{calculate_and_insert_percent_change}
\title{precent change and insertion}
\usage{
calculate_and_insert_percent_change(data1, data2, col_name)
}
\arguments{
\item{data1}{your original dataframe}

\item{data2}{your second dataframe}

\item{col_name}{the name of the column to calculate percent change for}
}
\value{
a dataframe
}
\description{
You need two data frames of equal size. The function assumes column 1 is an unused ID column.
}
\note{
for (col_name in colnames(data2)\link{-1}) {
data2 <- calculate_and_insert_percent_change(data1, data2, col_name)}
}
