% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{calculate_fstat_from_beta_se}
\alias{calculate_fstat_from_beta_se}
\title{calculate fstat from beta se}
\usage{
calculate_fstat_from_beta_se(data)
}
\arguments{
\item{data}{Data frame containing SNP data.}
}
\value{
Data frame with fstat values added.
}
\description{
calculate fstat from beta se
}
