% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{calculate_fstat_from_r2}
\alias{calculate_fstat_from_r2}
\title{calculate fstat from r2}
\usage{
calculate_fstat_from_r2(data, grouping_column)
}
\arguments{
\item{data}{Data frame containing SNP data.}

\item{grouping_column}{Column to group by.}
}
\value{
Data frame with fstat values added.
}
\description{
calculate fstat from r2
}
