% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{calculate_ICC}
\alias{calculate_ICC}
\title{Calculate Intraclass Correlation Coefficient (ICC)}
\usage{
calculate_ICC(df, id_samples, id_qc)
}
\arguments{
\item{df}{A data frame where rows = features and columns = samples
column names should be sample IDs; row names should be feature IDs}

\item{id_samples}{A vector of column names or indices representing the sample columns in \code{df}.}

\item{id_qc}{A vector of column names or indices representing the quality control columns in \code{df}.}
}
\value{
A named numeric vector of ICC values for each row in \code{df}.
}
\description{
This function calculates the Intraclass Correlation Coefficient (ICC) for a set of samples and quality controls
using a linear mixed-effects model.
From: https://pmc.ncbi.nlm.nih.gov/articles/PMC6570933/
Based on: https://github.com/courtneyschiffman/Metabolomics-Filtering/blob/master/ICC.R
}
