% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{calculate_r2}
\alias{calculate_r2}
\title{calculate r2}
\usage{
calculate_r2(data)
}
\arguments{
\item{data}{Data frame containing SNP data.}
}
\value{
Data frame with r2 values added.
}
\description{
calculate r2
}
