% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{check_palindromic}
\alias{check_palindromic}
\title{Checks if alleles are palindromic.}
\usage{
check_palindromic(A1, A2)
}
\arguments{
\item{A1}{The first allele.}

\item{A2}{The second allele.}
}
\value{
TRUE if alleles are palindromic, FALSE otherwise.
}
\description{
This function checks if the alleles provided are palindromic pairs (e.g., A-T, C-G).
}
