% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_check_alignment}
\alias{coloc_check_alignment}
\title{Check Alignment for Colocalisation}
\usage{
coloc_check_alignment(D, thr = 0.2, do_plot = TRUE)
}
\arguments{
\item{D}{Dataset containing beta, varbeta, and LD.}

\item{thr}{Threshold for alignment check.}

\item{do_plot}{Logical indicating whether to generate a plot.}
}
\value{
Mean of values greater than zero, indicating alignment quality.
}
\description{
This function checks the alignment for colocalisation based on the product of Z scores and LD.
}
