% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_manh.plot}
\alias{coloc_manh.plot}
\title{Plot Manhattan Plot for Colocalisation}
\usage{
coloc_manh.plot(
  df,
  wh,
  position = if ("position" \%in\% names(df)) {
     df$position
 } else {
    
    1:nrow(df)
 }
)
}
\arguments{
\item{df}{Data frame containing colocalisation data.}

\item{wh}{Which trait to plot (1 or 2).}

\item{position}{Chromosome positions or SNP numbers.}
}
\description{
Generates a Manhattan plot based on colocalisation data.
}
