% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_perform_colocalization}
\alias{coloc_perform_colocalization}
\title{Performs colocalization analysis with varying prior probabilities and saves the results}
\usage{
coloc_perform_colocalization(
  coloc_data_exposure,
  coloc_data_outcome,
  label,
  outcome,
  FILE_NAME,
  window
)
}
\arguments{
\item{coloc_data_exposure}{Colocalization data for the exposure trait.}

\item{coloc_data_outcome}{Colocalization data for the outcome trait.}

\item{label}{Label for the analysis.}

\item{outcome}{Outcome trait.}

\item{FILE_NAME}{Name of the file to save results.}

\item{window}{Window size for the analysis.}
}
\description{
Performs colocalization analysis with varying prior probabilities and saves the results
}
