% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_prior.adjust}
\alias{coloc_prior.adjust}
\title{Adjust Colocalisation Prior Probabilities}
\usage{
coloc_prior.adjust(summ, newp12, p1 = 1e-04, p2 = 1e-04, p12 = 1e-06)
}
\arguments{
\item{summ}{Summary vector or list containing necessary data.}

\item{newp12}{New p12 values for adjustment.}

\item{p1}{Probability p1.}

\item{p2}{Probability p2.}

\item{p12}{Probability p12.}
}
\value{
Adjusted posterior probabilities matrix.
}
\description{
Adjusts colocalisation prior probabilities based on new p12 values.
}
