% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_prior.snp2hyp}
\alias{coloc_prior.snp2hyp}
\title{Convert SNP Prior to Hypotheses}
\usage{
coloc_prior.snp2hyp(nsnp, p12 = 1e-06, p1 = 1e-04, p2 = 1e-04)
}
\arguments{
\item{nsnp}{Number of SNPs.}

\item{p12}{Probability p12.}

\item{p1}{Probability p1.}

\item{p2}{Probability p2.}
}
\value{
Matrix of hypotheses probabilities.
}
\description{
Converts SNP prior probabilities to hypotheses probabilities for colocalisation.
}
