% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{coloc_sensitivity}
\alias{coloc_sensitivity}
\title{Plots sensitivity analysis for coloc results}
\usage{
coloc_sensitivity(
  obj,
  rule = "H4 > 0.8",
  trait1_title = "trait 1",
  trait2_title = "trait 2",
  dataset1 = NULL,
  dataset2 = NULL,
  npoints = 100,
  suppress_messages = FALSE,
  row = 1,
  data_check_trait1 = NULL,
  data_check_trait2 = NULL
)
}
\arguments{
\item{obj}{Output from \code{coloc.detail} or \code{coloc.process}.}

\item{rule}{Decision rule for acceptable posterior probabilities.}

\item{trait1_title}{Title for the first trait (default: "trait 1").}

\item{trait2_title}{Title for the second trait (default: "trait 2").}

\item{dataset1}{Optional dataset for the first trait.}

\item{dataset2}{Optional dataset for the second trait.}

\item{npoints}{Number of points for evaluating prior values. Default is \code{100}.}

\item{suppress_messages}{Logical, whether to suppress messages (default is FALSE).}

\item{row}{Row number for coloc summary if multiple rows are present. Default is \code{1}.}

\item{data_check_trait1}{Data for trait 1 to be checked.}

\item{data_check_trait2}{Data for trait 2 to be checked.}
}
\value{
List of prior matrix, posterior matrix, and a pass/fail indicator (returned invisibly).
}
\description{
Plots sensitivity analysis for coloc results
}
