% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR-edits.R
\name{create_ids}
\alias{create_ids}
\title{Create IDs from input data}
\usage{
create_ids(x)
}
\arguments{
\item{x}{Input data.}
}
\value{
A vector of IDs.
}
\description{
This function creates IDs from the input data by converting the input to factors and assigning random strings as levels.
}
