% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_consistency.R
\name{directional_consistency}
\alias{directional_consistency}
\title{Workout consistency of effect estimates}
\usage{
directional_consistency(data)
}
\arguments{
\item{data}{your data frame}
}
\description{
Before using, your data needs to be in wide format, e.g., rows are exposures and columns are the betas for those IDs you want to check consistency for
Convert to wide: \code{data <- as.data.frame(pivot_wider(data, names_from = group, values_from = beta))}
}
