% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Perform enrichment analysis}
\usage{
enrichment_analysis(
  data,
  sig_direction,
  group_across,
  group_within,
  universe_geneinfo,
  database,
  minGSSize,
  maxGSSize,
  pvalueCutoff,
  qvalueCutoff,
  file_path
)
}
\arguments{
\item{data}{A data frame containing the data for analysis.}

\item{sig_direction}{The direction of significance (e.g., "up" or "down").}

\item{group_across}{The column name representing the grouping variable across which enrichment analysis is performed.}

\item{group_within}{The column name representing the grouping variable within which enrichment analysis is performed.}

\item{universe_geneinfo}{A vector containing gene information for the entire universe.}

\item{database}{A data frame containing the database for enrichment analysis.}

\item{minGSSize}{The minimum gene set size.}

\item{maxGSSize}{The maximum gene set size.}

\item{pvalueCutoff}{The p-value cutoff for significance.}

\item{qvalueCutoff}{The q-value cutoff for significance.}

\item{file_path}{The file path where results and plots will be saved.}
}
\description{
This function conducts enrichment analysis on the provided data using the enricher function.
}
