% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_analysis}
\alias{epic_somalogic_analysis}
\title{\code{epic_somalogic_analysis()}: Perform analysis on EPIC somalogic data}
\usage{
epic_somalogic_analysis(list_processed)
}
\arguments{
\item{list_processed}{A list from \code{epic_somalogic_process_data()}}
}
\value{
A data frame containing the results of Cox proportional hazards analysis
}
\description{
This function takes a processed list from \code{epic_somalogic_process_data()} and
performs Cox proportional hazards analysis using predefined models from
\code{epic_somalogic_models()}. It iterates over each combination of data frames
and exposures, fits the models, and stores the results in a data frame.
It prints the number of unique exposures and data frame combinations for sanity
}
