% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_create_analysis}
\alias{epic_somalogic_create_analysis}
\title{\code{epic_somalogic_create_analysis()}: process EPIC somalogic data for analysis}
\usage{
epic_somalogic_create_analysis(list_df)
}
\arguments{
\item{list_df}{A nested list containing EPIC somalogic data frames}
}
\value{
A processed nested list of EPIC somalogic data frames for analysis
}
\description{
This function preprocesses EPIC somalogic data for analysis by applying
Prentice weighting using \code{epic_somalogic_process_data()}.
}
