% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_create_sex}
\alias{epic_somalogic_create_sex}
\title{\code{epic_somalogic_create_sex()}: create sex-specific dataframes for a given 'ID_col' group for Epic SomaLogic data}
\usage{
epic_somalogic_create_sex(df)
}
\arguments{
\item{df}{A df with column "sex"}
}
\value{
A list containing sex-specific dataframes.
}
\description{
This function takes a subset of data for a specific 'ID_col' group and creates
sex-specific dataframes - sex-combined, male, and female.  Must use \code{lapply(list_data, epic_somalogic_create_sex)}
if you want to run this over a list of dataframes.
}
