% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_create_sex_specific_dataframes}
\alias{epic_somalogic_create_sex_specific_dataframes}
\title{Create sex-specific dataframes for a given 'ID_col' group for Epic SomaLogic data}
\usage{
epic_somalogic_create_sex_specific_dataframes(data_group, ID_col)
}
\arguments{
\item{data_group}{A subset of the original dataframe for a specific 'ID_col' group.}

\item{ID_col}{The unique identifier for the 'ID_col' group.}
}
\value{
A list containing sex-specific dataframes.
}
\description{
This function takes a subset of data for a specific 'ID_col' group and creates
sex-specific dataframes - sex-combined, male, and female.
}
