% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_create_subtypes}
\alias{epic_somalogic_create_subtypes}
\title{\code{epic_somalogic_create_subtypes()}: Create a list of dataframes with sex-specific and subtype data for Epic SomaLogic data.}
\usage{
epic_somalogic_create_subtypes(df)
}
\arguments{
\item{df}{A dataframe containing the Epic SomaLogic data.}
}
\value{
A list of dataframes with sex-specific and subtype data.
}
\description{
This function is a master function for 'epic_somalogic_create_sex_specific_dataframes()'
and 'epic_somalogic_create_subtypes()'. It takes a dataframe, splits it by
'ID_col', and creates a list of sex-specific and subtype dataframes
for each 'ID_col' group. 'ID_col' is the filename (e.g., cancer-transformed.txt)
}
