% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_exclusions}
\alias{epic_somalogic_exclusions}
\title{\code{epic_somalogic_exclusions()}: process exclusions from \code{epic_somalogic_create_subtype_dataframes()}}
\usage{
epic_somalogic_exclusions(list_data, exclusions)
}
\arguments{
\item{list_data}{A list of data frames.}

\item{exclusions}{A list of exclusions for each data frame.}
}
\value{
The modified list_data after applying exclusions.
}
\description{
This function takes a nested list of data frames and a nested list of exclusions, and applies
the exclusion logic to remove specified rows (samples) and columns (features) from the data frames.
}
