% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_format_data}
\alias{epic_somalogic_format_data}
\title{\code{epic_somalogic_format_data()}: process standard variables}
\usage{
epic_somalogic_format_data(df)
}
\arguments{
\item{df}{A dataframe containing EPIC-SOMALOGIC data.}
}
\value{
A processed dataframe with calculated follow-up time, converted categorical variables, and reordered factor levels.
}
\description{
This function processes EPIC-SOMALOGIC data by calculating follow-up time,
converting certain categorical variables to numeric, and reordering factor levels.
}
\details{
The function calculates the follow-up time by subtracting the age at cancer onset
from the current age. It converts the "cvd_t2d_coh" variable to numeric, where "No"
is converted to 0 and "Yes" to 1. It creates age groups with a 5-year interval
and reorders factor levels for variables "smoke_stat", "pa_index", and "l_school".
}
