% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_models}
\alias{epic_somalogic_models}
\title{\code{epic_somalogic_models()}: a list of coxph models}
\usage{
epic_somalogic_models()
}
\value{
A list containing three Cox proportional hazards models
}
\description{
This function defines three Cox proportional hazards models including
}
