% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_prentic_weighting}
\alias{epic_somalogic_prentic_weighting}
\title{\code{epic_somalogic_prentic_weighting()}: Prentice weighting function for EPIC somalogic analysis}
\usage{
epic_somalogic_prentic_weighting(data, followup_years)
}
\arguments{
\item{data}{A data frame containing the EPIC somalogic data}

\item{followup_years}{The number of follow-up years used for weighting}
}
\value{
A processed data frame with Prentice weighting applied
}
\description{
This function applies Prentice weighting to the provided data based on follow-up years.
It filters out cases where the age at event is less than the age plus follow-up years,
and then creates different subsets of the data based on disease status and cohort membership.
}
