% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_process_data}
\alias{epic_somalogic_process_data}
\title{\code{epic_somalogic_process_data()}: creates nested raw and complete dataframes}
\usage{
epic_somalogic_process_data(list_df)
}
\arguments{
\item{list_df}{A nested list containing EPIC-SOMALOGIC dataframes.}
}
\value{
A nested list of processed EPIC-SOMALOGIC dataframes with calculated follow-up time,
converted categorical variables, and reordered factor levels.
}
\description{
the function applies \code{epic_somalogic_create_datacomplete()} to a nested list and creates
a nested list of raw and complete data frames
}
