% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_somalogic_data.R
\name{epic_somalogic_process_data_complete}
\alias{epic_somalogic_process_data_complete}
\title{\code{epic_somalogic_process_data_complete()}: process missing values in EPIC somalogic data}
\usage{
epic_somalogic_process_data_complete(df_complete, followup_years)
}
\arguments{
\item{df_complete}{A complete data frame containing EPIC somalogic data}

\item{followup_years}{The number of follow-up years for the analysis}
}
\value{
A processed data frame with missing values replaced by -9
}
\description{
This function converts missing values in EPIC somalogic data to -9 and
calculates the percentage of missing values for each variable.
}
