% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{filter_case_control}
\alias{filter_case_control}
\title{Filter Samples Based on Case Control Criteria}
\usage{
filter_case_control(df, df_samples, col_case_control, col_samples)
}
\arguments{
\item{df}{A data frame containing feature data where rows represent samples and columns represent features.}

\item{df_samples}{A data frame containing sample metadata with at least the column specified by \code{col_case_control}.}

\item{col_case_control}{A string specifying the column name in \code{df_samples} that indicates case control groups.}

\item{col_samples}{A string specifying the column name in \code{df_samples} to use for sample filtering (e.g., ID column).}
}
\value{
A list containing:
\itemize{
\item \code{filtered_df}: The filtered feature data frame.
\item \code{filtered_samples}: The filtered sample data frame.
\item \code{excluded_ids}: A vector of IDs of excluded samples.
}
}
\description{
This function filters the samples in the provided data frame based on case control criteria.
It excludes individuals without a matched caseset and those matched casesets with
more than two individuals. It also provides information on excluded samples.
}
