% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{finemap_pval_LD}
\alias{finemap_pval_LD}
\title{Perform LD clumping on a dataframe based on p-value threshold}
\usage{
finemap_pval_LD(df, bfile, pval_threshold = 5e-08, clump_r2 = 0.001)
}
\arguments{
\item{df}{A data frame containing genetic association results with at least the following columns: \code{pval} (p-value) and \code{rsid} (SNP identifier).}

\item{bfile}{A character string specifying the prefix for the PLINK binary files (e.g., \code{.bed}, \code{.bim}, \code{.fam}).}

\item{pval_threshold}{A numeric value specifying the p-value threshold for filtering SNPs. Default is \code{5E-8}.}

\item{clump_r2}{A numeric vector specifying the r2 threshold(s) for LD clumping. Default is \code{0.001}. The function will perform clumping for each r2 value.}
}
\value{
A tibble with columns \code{SNP}, \code{test}, and \code{LD}, where:
\itemize{
\item \code{SNP}: the SNPs selected after clumping.
\item \code{test}: a character string indicating the test, always "p_ld".
\item \code{LD}: the r2 threshold used for clumping.
}
}
\description{
This function performs linkage disequilibrium (LD) clumping on a given dataframe of genetic association results. The function filters the dataframe based on the provided p-value threshold and then applies the LD clumping for different values of \code{r2}. If clumping fails, the function returns a dataframe with \code{NA} values in the \code{SNP} and \code{test} columns.
}
\examples{
# Example usage
\dontrun{
result <- finemap_pval_LD(df = my_data, bfile = "path_to_bfile", pval_threshold = 5E-8, clump_r2 = c(0.01, 0.1))
}
}
