% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{finimom_cs_table}
\alias{finimom_cs_table}
\title{Create a Table of Credible Sets from Finimom Model}
\usage{
finimom_cs_table(finimom_model, df)
}
\arguments{
\item{finimom_model}{A Finimom model object containing credible sets and PIPs.}

\item{df}{A data frame containing SNP information with columns:
\describe{
\item{SNP}{Character vector of SNP identifiers.}
\item{POS}{Numeric vector of SNP positions.}
\item{P}{Numeric vector of p-values.}
}}
}
\value{
A data frame summarizing the credible sets with the following columns:
\itemize{
\item SNP: SNP identifier.
\item POS: SNP position.
\item P: SNP p-value.
\item PIP: Posterior inclusion probability for the SNP.
\item cs: Credible set label (e.g., "L1", "L2").
\item cs_snps: Comma-separated list of other SNPs in the same credible set.
\item label: SNP identifier of the SNP with the smallest p-value in each credible set. We choose P because PIP is equal for SNPs in a CS.
}
}
\description{
This function processes a Finimom model to generate a table summarizing credible sets,
their associated SNPs, positions, and posterior inclusion probabilities (PIPs).
Additionally, it identifies the SNP with the smallest p-value in each credible set.
}
