% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{flip_alleles}
\alias{flip_alleles}
\title{Flips alleles.}
\usage{
flip_alleles(x)
}
\arguments{
\item{x}{The allele to be flipped.}
}
\value{
The flipped allele.
}
\description{
This function flips alleles by converting lowercase letters to uppercase and vice versa.
}
