% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forestplot_label}
\alias{forestplot_label}
\title{Forest Plot with Labels}
\usage{
forestplot_label(
  df,
  name,
  estimate,
  se,
  pvalue = NULL,
  label_column = NULL,
  label_width = 1,
  colour = NULL,
  shape = NULL,
  logodds = FALSE,
  psignif = 0.05,
  ci = 0.95,
  ...
)
}
\arguments{
\item{df}{A data frame with the data to plot. It must contain at least three
variables: a character column with the names to be displayed on the y-axis
(see parameter \code{name}), a numeric column with the value (or the log of the
value) to display (see parameter \code{estimate}), and a numeric value with
the corresponding standard errors (see parameter \code{se}). It may contain
additional columns, e.g., the corresponding p-values (see parameter \code{pvalue}),
in conjunction with the threshold given in \code{psignif}; non-significant results
will be displayed as hollow points. Other variables may be used as ggplot2::aesthetics to define
the colour and the shape of the points to be plotted.}

\item{name}{The variable in \code{df} that contains the y-axis
names. This argument is automatically \link[rlang:quotation]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the \code{df} data frame.}

\item{estimate}{The variable in \code{df} that contains the values (or log of
values) to be displayed. This argument is automatically
\link[rlang:quotation]{quoted} and \link[rlang:eval_tidy]{evaluated} in the
context of the \code{df} data frame.}

\item{se}{The variable in the \code{df} data frame that contains the standard
error values. This argument is automatically \link[rlang:quotation]{quoted}
and \link[rlang:eval_tidy]{evaluated} in the context of the \code{df} data
frame.}

\item{pvalue}{The variable in \code{df} that contains the
p-values. Defaults to NULL. When explicitly defined, in conjunction with
the p-value threshold provided in the \code{psignif}, non-significant
entries will be drawn as hollow points.}

\item{label_column}{The variable in \code{df} that contains the labels to display.
This argument is automatically \link[rlang:quotation]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the \code{df} data frame.}

\item{label_width}{A numeric value to specify the relative width of the label plot
in the combined output. Adjust this to control the width of the labels compared
to the forest plot.}

\item{colour}{The variable in \code{df} by which to colour the different
groups of points.}

\item{shape}{The variable in \code{df} by which to shape the different groups of
points.}

\item{logodds}{Logical (defaults to FALSE) specifying whether the \code{estimate}
parameter should be treated as log odds/hazards ratio (TRUE) or not (FALSE).}

\item{psignif}{Numeric, defaults to 0.05. The p-value threshold
for statistical significance.}

\item{ci}{A number between 0 and 1 (defaults to 0.95) indicating the type of
confidence interval to be drawn.}

\item{...}{\code{ggplot2} graphical parameters such as \code{title},
\code{ylab}, \code{xlab}, \code{xtickbreaks}, etc. to be passed along.}
}
\value{
A \code{ggplot} object.
}
\description{
Forest Plot with Labels
}
\note{
See \code{vignette(programming, package = "dplyr")} for an
introduction to non-standard evaluation.
}
