% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR-edits.R
\name{format_data}
\alias{format_data}
\title{Format data for Mendelian Randomization (MR) analysis}
\usage{
format_data(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{Input data frame.}

\item{type}{Type of data (e.g., "exposure", "outcome").}

\item{snps}{Vector of SNP IDs to subset data.}

\item{header}{Logical indicating if the input data has headers.}

\item{phenotype_col}{Name of the phenotype column.}

\item{snp_col}{Name of the SNP column.}

\item{beta_col}{Name of the beta column.}

\item{se_col}{Name of the standard error column.}

\item{eaf_col}{Name of the effect allele frequency column.}

\item{effect_allele_col}{Name of the effect allele column.}

\item{other_allele_col}{Name of the other allele column.}

\item{pval_col}{Name of the p-value column.}

\item{units_col}{Name of the units column.}

\item{ncase_col}{Name of the case count column.}

\item{ncontrol_col}{Name of the control count column.}

\item{samplesize_col}{Name of the sample size column.}

\item{gene_col}{Name of the gene column.}

\item{id_col}{Name of the ID column.}

\item{min_pval}{Minimum p-value threshold.}

\item{z_col}{Name of the z-score column.}

\item{info_col}{Name of the INFO column.}

\item{chr_col}{Name of the chromosome column.}

\item{pos_col}{Name of the position column.}

\item{log_pval}{Logical indicating if the p-value should be log-transformed.}
}
\value{
Formatted data frame for MR analysis.
}
\description{
This function formats input data for Mendelian Randomization (MR) analysis, ensuring required columns are present and have appropriate data types.
}
