% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{format_data2}
\alias{format_data2}
\title{Format Data Function}
\usage{
format_data2(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{The dataset to be formatted.}

\item{type}{Character string specifying the type of data, defaults to "exposure".}

\item{snps}{Vector of SNP IDs to subset the data.}

\item{header}{Logical indicating if the dataset has column names, defaults to TRUE.}

\item{phenotype_col}{Character string specifying the column name for phenotype data, defaults to "Phenotype".}

\item{snp_col}{Character string specifying the column name for SNP IDs, defaults to "SNP".}

\item{beta_col}{Character string specifying the column name for beta coefficients, defaults to "beta".}

\item{se_col}{Character string specifying the column name for standard errors, defaults to "se".}

\item{eaf_col}{Character string specifying the column name for effect allele frequency, defaults to "eaf".}

\item{effect_allele_col}{Character string specifying the column name for effect allele, defaults to "effect_allele".}

\item{other_allele_col}{Character string specifying the column name for other allele, defaults to "other_allele".}

\item{pval_col}{Character string specifying the column name for p-values, defaults to "pval".}

\item{units_col}{Character string specifying the column name for units, defaults to "units".}

\item{ncase_col}{Character string specifying the column name for number of cases, defaults to "ncase".}

\item{ncontrol_col}{Character string specifying the column name for number of controls, defaults to "ncontrol".}

\item{samplesize_col}{Character string specifying the column name for sample size, defaults to "samplesize".}

\item{gene_col}{Character string specifying the column name for gene, defaults to "gene".}

\item{id_col}{Character string specifying the column name for IDs, defaults to "id".}

\item{min_pval}{Minimum p-value to replace non-finite or extreme p-values, defaults to 1e-200.}

\item{z_col}{Character string specifying the column name for z-scores, defaults to "z".}

\item{info_col}{Character string specifying the column name for info scores, defaults to "info".}

\item{chr_col}{Character string specifying the column name for chromosome, defaults to "chr".}

\item{pos_col}{Character string specifying the column name for position, defaults to "pos".}

\item{log_pval}{Logical indicating if p-values should be log-transformed, defaults to FALSE.}
}
\value{
A formatted dataset ready for analysis.
}
\description{
This function formats a dataset for analysis, particularly for Mendelian Randomization (MR) studies.
}
