% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise}
\alias{harmonise}
\title{Harmonises alleles for various allele configurations.}
\usage{
harmonise(dat, tolerance, action)
}
\arguments{
\item{dat}{Data frame containing SNP information.}

\item{tolerance}{Tolerance for allele frequency matching.}

\item{action}{Action to take for ambiguous or incompatible alleles.}
}
\value{
A data frame with harmonised alleles.
}
\description{
This function harmonises alleles for various allele configurations.
}
