% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise_11}
\alias{harmonise_11}
\title{Harmonises alleles for a 1-1 allele configuration.}
\usage{
harmonise_11(SNP, A1, B1, betaA, betaB, fA, fB, tolerance, action)
}
\arguments{
\item{SNP}{Vector of SNP IDs.}

\item{A1}{First allele for exposure.}

\item{B1}{First allele for outcome.}

\item{betaA}{Effect sizes for exposure.}

\item{betaB}{Effect sizes for outcome.}

\item{fA}{Allele frequencies for exposure.}

\item{fB}{Allele frequencies for outcome.}

\item{tolerance}{Tolerance for allele frequency matching.}

\item{action}{Action to take for ambiguous or incompatible alleles.}
}
\value{
A data frame with harmonised alleles.
}
\description{
This function harmonises alleles for a 1-1 allele configuration.
}
