% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise_22}
\alias{harmonise_22}
\title{Harmonises alleles for a 2-2 allele configuration.}
\usage{
harmonise_22(SNP, A1, A2, B1, B2, betaA, betaB, fA, fB, tolerance, action)
}
\arguments{
\item{SNP}{Vector of SNP identifiers.}

\item{A1}{First allele for exposure.}

\item{A2}{Second allele for exposure.}

\item{B1}{First allele for outcome.}

\item{B2}{Second allele for outcome.}

\item{betaA}{Beta coefficient for exposure.}

\item{betaB}{Beta coefficient for outcome.}

\item{fA}{Frequency of the first allele for exposure.}

\item{fB}{Frequency of the second allele for exposure.}

\item{tolerance}{Tolerance level for allele frequency.}

\item{action}{Action to take for ambiguous alleles.}
}
\value{
A data frame with harmonised alleles for exposure and outcome.
}
\description{
This function harmonises alleles for a 2-2 allele configuration by flipping and recoding indels.
}
