% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise_cleanup_variables}
\alias{harmonise_cleanup_variables}
\title{Cleans up variables in the result table.}
\usage{
harmonise_cleanup_variables(res.tab)
}
\arguments{
\item{res.tab}{The result table to be cleaned up.}
}
\value{
The cleaned-up result table.
}
\description{
This function cleans up variables in the result table to ensure they are in the appropriate format.
Specifically, it converts certain columns to numeric format and ensures alleles are in uppercase.
}
