% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise_data}
\alias{harmonise_data}
\title{\code{harmonise_data()}: modification of \code{TwoSampleMR::harmonise_data()}}
\usage{
harmonise_data(exposure_dat, outcome_dat, action = 2)
}
\arguments{
\item{exposure_dat}{Output from \code{\link[=read_exposure_data]{read_exposure_data()}}.}

\item{outcome_dat}{Output from \code{\link[=extract_outcome_data]{extract_outcome_data()}}.}

\item{action}{Level of strictness in dealing with SNPs.
\itemize{
\item \code{action = 1}: Assume all alleles are coded on the forward strand, i.e. do not attempt to flip alleles
\item \code{action = 2}: Try to infer positive strand alleles, using allele frequencies for palindromes (default, conservative);
\item \code{action = 3}: Correct strand for non-palindromic SNPs, and drop all palindromic SNPs from the analysis (more conservative).
If a single value is passed then this action is applied to all outcomes.
But multiple values can be supplied as a vector, each element relating to a different outcome.
}}
}
\description{
\code{harmonise_data()}: modification of \code{TwoSampleMR::harmonise_data()}
}
\details{
The code is modified at line 6  where \code{res.tab} is first created. I used left_join
and then remove any rows where there is NA in column id.outcome.
This is because when doing mltiple exposures and multiple outcomes you can have
situations where the exposure SNP isnt present in the outcome but a proxy SNP is
already an exposure SNP so you end up with duplicate SNPs. This fixes that issue.
Everything else is the same.
}
