% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_sex-specific.R
\name{harmonise_data_sex}
\alias{harmonise_data_sex}
\title{harmonise_data_sex: perform harmonise_data() using sex specific data}
\usage{
harmonise_data_sex(data_exposure, data_outcome, harmonise_action)
}
\arguments{
\item{data_exposure}{exposure data frame}

\item{data_outcome}{outcome data frame}

\item{harmonise_action}{action for harmonising alleles (1,2,3)}
}
\value{
a data frame
}
\description{
harmonise_data_sex: perform harmonise_data() using sex specific data
}
\note{
you must have a single column called 'sex' in your exposure and outcome data
}
