% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{harmonise_make_snp_effects_positive}
\alias{harmonise_make_snp_effects_positive}
\title{Makes SNP effects positive.}
\usage{
harmonise_make_snp_effects_positive(res.tab)
}
\arguments{
\item{res.tab}{The result table with SNP information.}
}
\value{
The result table with SNP effects made positive.
}
\description{
This function ensures that SNP effects are positive by adjusting allele frequencies and beta coefficients accordingly.
It identifies SNPs with negative beta coefficients and swaps alleles accordingly.
}
