% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_test.R
\name{imputation_test}
\alias{imputation_test}
\title{Perform Imputation Test}
\usage{
imputation_test(data, output_dir, subtitle)
}
\arguments{
\item{data}{your data variable e.g., mtcars}

\item{output_dir}{where you want to save the output}

\item{subtitle}{what you want to call this imputation test}
}
\value{
The result of the imputation test.
}
\description{
This function performs an imputation test using the specified qmd file.
Your data should have samples as rows and features as columns. There should
be no extra columns.
}
\examples{
\dontrun{
imputation_test(rmd_location = "rmd/imputation-test.Rmd",
                file_path = "project1/data/",
                data_var = "mtcars")
}

}
