% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscomparer_functions.R
\name{locuscomparer_add_label}
\alias{locuscomparer_add_label}
\title{Add Labels to df Data Frame}
\usage{
locuscomparer_add_label(df, snp)
}
\arguments{
\item{df}{(data.frame) A data frame containing at least the column \code{rsid}.}

\item{snp}{(character vector) A vector of SNP rsIDs to be labeled.}
}
\value{
(data.frame) The input data frame with an additional column \code{label} that contains the rsIDs
of the specified SNPs and empty strings for others.
}
\description{
This function adds labels to a df data frame based on the presence of specified SNPs.
}
