% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscomparer_functions.R
\name{locuscomparer_make_combined_plot}
\alias{locuscomparer_make_combined_plot}
\title{Generate a combined plot with two LocusZoom plots and a LocusCompare plot using coloc data.}
\usage{
locuscomparer_make_combined_plot(
  df,
  trait1_title = "exposure",
  trait2_title = "outcome",
  ld,
  colour = colour,
  shape = shape,
  size = size,
  snp = NULL,
  combine = TRUE,
  legend = TRUE,
  legend_position = c("bottomright", "topright", "topleft")
)
}
\arguments{
\item{df}{(data.frame) An input data.frame containing the following columns:
\itemize{
\item \code{rsid} (character) - SNP rsID.
\item \code{pval1} (numeric) - p-value for study 1.
\item \code{logp1} (numeric) - -log10(p-value) for study 1.
\item \code{logp2} (numeric) - -log10(p-value) for study 2.
\item \code{chr} (character) - Chromosome name (without 'chr').
\item \code{pos} (integer) - Position on the chromosome.
\item \code{label} (character) - Optional label for SNPs.
}
See the example for \code{get_lead_snp()} for details on how to generate this data.frame.}

\item{trait1_title}{(string) Title for the x-axis of the LocusCompare plot, representing the first trait (exposure). Default is \code{"exposure"}.}

\item{trait2_title}{(string) Title for the y-axis of the LocusCompare plot, representing the second trait (outcome). Default is \code{"outcome"}.}

\item{ld}{(data.frame) The output from \code{retrieve_LD()}. This data.frame should contain columns:
\itemize{
\item \code{SNP_A} (character) - SNP rsID for the reference SNP.
\item \code{SNP_B} (character) - SNP rsID for the comparison SNP.
\item \code{R2} (numeric) - LD (R^2) value between the reference and comparison SNPs.
}}

\item{colour}{(named vector) A vector specifying the colors for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = "#FF0000", "rs2" = "#0000FF", "rs3" = "#00FF00")}.}

\item{shape}{(named vector) A vector specifying the shapes for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = 21, "rs2" = 22, "rs3" = 23)}.}

\item{size}{(named vector) A vector specifying the sizes for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = 2, "rs2" = 3, "rs3" = 4)}.}

\item{snp}{(string, optional) SNP rsID to be highlighted. If NULL, the function will select the lead SNP. Default: NULL.}

\item{combine}{(boolean, optional) Should the three plots be combined into one plot? If FALSE, a list of three plots will be returned. Default: TRUE.}

\item{legend}{(boolean, optional) Should the legend be shown in the LocusCompare plot? Default: TRUE.}

\item{legend_position}{(string, optional) Position of the legend in the LocusCompare plot. Choose from 'bottomright', 'topright', or 'topleft'. Default: 'bottomright'.}
}
\value{
(ggplot) A combined plot with a LocusCompare plot and two LocusZoom plots, or a list of the three individual plots if \code{combine} is set to FALSE.
}
\description{
This function creates a combined plot featuring a LocusCompare plot alongside two LocusZoom plots. Each LocusZoom plot represents an association study for a specific trait.
The LocusCompare plot compares the -log10(p-values) of two traits, while the LocusZoom plots show the detailed locus zoom views for the two traits.
}
