% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscomparer_functions.R
\name{locuscomparer_make_plot}
\alias{locuscomparer_make_plot}
\title{Generate a combined plot with two LocusZoom plots and a LocusCompare plot using coloc data}
\usage{
locuscomparer_make_plot(
  data_coloc_exposure,
  data_coloc_outcome,
  SNP_causal_exposure,
  trait1_title = "exposure",
  trait2_title = "outcome"
)
}
\arguments{
\item{data_coloc_exposure}{(data.frame) Data frame containing the exposure data with the following columns:
\itemize{
\item \code{snp} (character) - SNP rsID.
\item \code{pval} (numeric) - p-value for the association.
\item \code{chr} (character) - Chromosome name.
\item \code{position} (integer) - Position on the chromosome.
\item \code{LD} (matrix) - LD matrix for the SNPs.
}}

\item{data_coloc_outcome}{(data.frame) Data frame containing the outcome data with the following columns:
\itemize{
\item \code{snp} (character) - SNP rsID.
\item \code{pval} (numeric) - p-value for the association.
}}

\item{SNP_causal_exposure}{(character vector) A vector of SNP rsIDs to be labeled.}

\item{trait1_title}{(character) The title for the first trait (exposure). Default is "exposure".}

\item{trait2_title}{(character) The title for the second trait (outcome). Default is "outcome".}
}
\value{
(ggplot) Combined plot of the two LocusZoom plots and the LocusCompare plot.
}
\description{
This function creates a combined plot with two LocusZoom plots and a LocusCompare plot. Each LocusZoom plot represents an association study.
}
