% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscomparer_functions.R
\name{locuscomparer_make_scatterplot}
\alias{locuscomparer_make_scatterplot}
\title{Create a Scatter Plot Comparing Two Traits}
\usage{
locuscomparer_make_scatterplot(
  df,
  trait1_title,
  trait2_title,
  colour,
  shape,
  size,
  legend = TRUE,
  legend_position = c("bottomright", "topright", "topleft")
)
}
\arguments{
\item{df}{(data.frame) A data frame containing the following columns:
\itemize{
\item \code{logp1} (numeric) - -log10(p-value) for the first trait.
\item \code{logp2} (numeric) - -log10(p-value) for the second trait.
\item \code{rsid} (character) - SNP rsID.
\item \code{label} (character) - Optional label for SNPs, used for annotation.
\item Additional columns for plotting aesthetics: \code{colour}, \code{shape}, \code{size} can be specified for different SNPs.
}
The \code{logp1} and \code{logp2} columns represent the transformed p-values for the two traits being compared.}

\item{trait1_title}{(string) The title for the x-axis of the scatter plot representing the first trait. Default is \code{"Trait 1"}.}

\item{trait2_title}{(string) The title for the y-axis of the scatter plot representing the second trait. Default is \code{"Trait 2"}.}

\item{colour}{(named vector) A vector specifying the colors for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = "#FF0000", "rs2" = "#0000FF", "rs3" = "#00FF00")}.}

\item{shape}{(named vector) A vector specifying the shapes for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = 21, "rs2" = 22, "rs3" = 23)}.}

\item{size}{(named vector) A vector specifying the sizes for different SNPs. The names of the vector should correspond to \code{rsid} values.
Example: \code{c("rs1" = 2, "rs2" = 3, "rs3" = 4)}.}

\item{legend}{(boolean) Whether to include a legend in the plot. If \code{TRUE}, a legend will be displayed showing the colors, shapes, and sizes associated with different SNPs. Default is \code{TRUE}.}

\item{legend_position}{(string) Position of the legend on the plot. Choose from 'bottomright', 'topright', or 'topleft'. Default is 'bottomright'.}
}
\value{
A \code{ggplot} object representing the scatter plot comparing two traits.
}
\description{
This function generates a scatter plot comparing the -log10(p-values) of two different traits. The plot includes points representing SNPs, with customizable aesthetics such as color, shape, and size based on the SNPs. Additionally, it can display a legend with configurable position.
}
