% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{missing_EAF}
\alias{missing_EAF}
\title{Fill Missing EAF Values in Data Frame}
\usage{
missing_EAF(df, reference, column_EAF)
}
\arguments{
\item{df}{The input dataframe containing the outcome data.}

\item{reference}{The file path to the reference data containing EAF information.}

\item{column_EAF}{The name of the column containing EAF values. If no column then this will be the name of the new column.}
}
\value{
The input dataframe with missing EAF values filled.
}
\description{
This function checks if the specified column containing EAF values in the input dataframe contains
any missing values (NA). If missing values are found, it reads the necessary
reference data to fill in the missing values based on the provided allele
information.
}
