% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_test.R
\name{model_test}
\alias{model_test}
\title{Provide akaike information criterion for all possible covariate combinations}
\usage{
model_test(data, exposure, outcome, covariates, match_ID = NA, model)
}
\arguments{
\item{data}{your data frame}

\item{exposure}{name of the exposure/predictor/independent variable column}

\item{outcome}{name of the outcome/dependent variable column}

\item{covariates}{name(s) of the covariate column(s)}

\item{match_ID}{name of the column with your case/control match ID}

\item{model}{name of the model you want to run: "lm" = stats::lm(); "clogit" = survival::clogit()}
}
\value{
an object from AICcmodavg::aictab()
}
\description{
Provide akaike information criterion for all possible covariate combinations
}
\examples{
model_test(data = my_data, exposure = "BMI", outcome = "breast_cancer", covariates = c("age", "sex"),  match_ID = "match_caseset", model = "clogit")

}
