% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions_sex-specific.R
\name{mr_heterogeneity_sex}
\alias{mr_heterogeneity_sex}
\title{mr_heterogeneity_sex: perform mr_heterogeneity() using sex specific data}
\usage{
mr_heterogeneity_sex(data_harmonise, methods_heterogeneity)
}
\arguments{
\item{data_harmonise}{harmonised data frame}

\item{methods_heterogeneity}{a list of methods to supply to method_list}
}
\value{
a data frame
}
\description{
mr_heterogeneity_sex: perform mr_heterogeneity() using sex specific data
}
\note{
you must have a single column called 'sex' in your harmonise data frame
}
