% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_sex-specific.R
\name{mr_leaveoneout_sex}
\alias{mr_leaveoneout_sex}
\title{mr_leaveoneout_sex: perform mr_leaveoneout() using sex specific data}
\usage{
mr_leaveoneout_sex(data_harmonise)
}
\arguments{
\item{data_harmonise}{harmonised data frame}
}
\value{
a data frame
}
\description{
mr_leaveoneout_sex: perform mr_leaveoneout() using sex specific data
}
\note{
you must have a single column called 'sex' in your harmonise data frame
}
