% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions_sex-specific.R
\name{mr_pleiotropy_test_sex}
\alias{mr_pleiotropy_test_sex}
\title{mr_pleiotropy_test_sex: perform mr_pleiotropy_test() using sex specific data}
\usage{
mr_pleiotropy_test_sex(data_harmonise)
}
\arguments{
\item{data_harmonise}{harmonised data frame}
}
\value{
a data frame
}
\description{
mr_pleiotropy_test_sex: perform mr_pleiotropy_test() using sex specific data
}
\note{
you must have a single column called 'sex' in your harmonise data frame
}
