% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_plots.R
\name{mr_scatter_plot}
\alias{mr_scatter_plot}
\title{modification of TwoSampleMR mr_scatter_plot() function that adds exposure/outcome labels}
\usage{
mr_scatter_plot(mr_results, dat)
}
\arguments{
\item{mr_results}{Output from mr().}

\item{dat}{Output from harmonise_data().}
}
\value{
List of plots
}
\description{
Requires dev version of ggplot2
}
