% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_sex-specific.R
\name{mr_sex}
\alias{mr_sex}
\title{mr_sex: perform mr() using sex specific data}
\usage{
mr_sex(data_harmonise, methods)
}
\arguments{
\item{data_harmonise}{hamronised data frame}

\item{methods}{a list of methds supplied to "method_list"}
}
\value{
a data frame
}
\description{
mr_sex: perform mr() using sex specific data
}
\note{
you must have a single column called 'sex' in your harmonised data frame
}
