% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions_sex-specific.R
\name{mr_singlesnp_sex}
\alias{mr_singlesnp_sex}
\title{mr_singlesnp_sex: perform mr_singlesnp() using sex specific data}
\usage{
mr_singlesnp_sex(data_harmonise)
}
\arguments{
\item{data_harmonise}{harmonised data frame}
}
\value{
a data frame
}
\description{
mr_singlesnp_sex: perform mr_singlesnp() using sex specific data
}
\note{
you must have a single column called 'sex' in your harmonise data frame
}
