% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_functions.R
\name{my_sensitivity}
\alias{my_sensitivity}
\title{Plots sensitivity analysis for coloc results}
\usage{
my_sensitivity(
  obj,
  rule = "",
  trait1_title,
  trait2_title,
  dataset1 = NULL,
  dataset2 = NULL,
  npoints = 100,
  doplot = TRUE,
  plot.manhattans = TRUE,
  preserve.par = FALSE,
  row = 1
)
}
\arguments{
\item{obj}{Output from \code{coloc.detail} or \code{coloc.process}.}

\item{rule}{Decision rule for acceptable posterior probabilities.}

\item{trait1_title}{Title for the first trait.}

\item{trait2_title}{Title for the second trait.}

\item{dataset1}{Optional dataset for the first trait.}

\item{dataset2}{Optional dataset for the second trait.}

\item{npoints}{Number of points for evaluating prior values. Default is \code{100}.}

\item{doplot}{Whether to draw the plot. Default is \code{TRUE}.}

\item{plot.manhattans}{Whether to show Manhattan plots. Default is \code{TRUE}.}

\item{preserve.par}{Whether to preserve the current graphics device parameters. Default is \code{FALSE}.}

\item{row}{Row number for coloc summary if multiple rows are present. Default is \code{1}.}
}
\value{
List of prior matrix, posterior matrix, and a pass/fail indicator (returned invisibly).
}
\description{
Plots sensitivity analysis for coloc results
}
