% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{normalization_residualMixedModels}
\alias{normalization_residualMixedModels}
\title{Normalize feature Data Using Residual Mixed Models}
\usage{
normalization_residualMixedModels(
  list,
  forIdentifier = c("Study", "Batch", "CaseSet", "Idepic"),
  listRandom = NULL,
  listFixedToKeep = NULL,
  listFixedToRemove = NULL,
  HeteroSked = NULL
)
}
\arguments{
\item{list}{A list containing the following elements:
\describe{
\item{data_features}{A data frame or tibble of dimensions n x (K+p) with:
\describe{
\item{n}{Number of observations.}
\item{p}{Number of features.}
\item{K}{Number of variables used for unique identification of individuals.}
}}
\item{data_samples}{A data frame or tibble of dimensions n x (K+d) with:
\describe{
\item{n}{Number of observations.}
\item{K}{Number of unique identifiers.}
\item{d}{Additional variables useful in final analysis (e.g., country, age, BMI).}
}}
\item{data_meta_features}{A p x 3 matrix indicating each feature's Name, Class, and Type.}
}}

\item{forIdentifier}{A character vector of strings indicating the names of variables used for unique identification of individuals.}

\item{listRandom}{A character vector of strings containing variable names modeled as random effects to be removed. If not NULL, should be either of length 1 or contain nested variables.}

\item{listFixedToKeep}{A character vector of strings containing variable names modeled as fixed effects to be kept.}

\item{listFixedToRemove}{A character vector of strings containing variable names modeled as fixed effects to be removed.}

\item{HeteroSked}{A string or NULL. If not NULL, the name of the variable for which heteroskedasticity will be accounted for. Must be included in \code{listRandom}.}
}
\value{
A list with:
\item{data}{A tibble with unwanted variation removed.}
\item{data_samples}{The input 'data_samples' data frame, ordered by IdentifierPipeline.}
\item{data_meta_features}{The input 'data_meta_features' matrix.}
}
\description{
This function normalizes feature data by removing unwanted effects using mixed models.
It accounts for random effects and fixed effects specified by the user, and optionally
corrects for heteroskedasticity. Residuals are calculated with \code{FUNnormalization_residualMixedModels()}.
The method is details in \url{https://www.mdpi.com/2218-1989/11/9/631}
}
