% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{outlier_pca_lof}
\alias{outlier_pca_lof}
\title{Perform PCA and Identify Outliers}
\usage{
outlier_pca_lof(df, col_samples, col_features, df_samples, df_features)
}
\arguments{
\item{df}{A data frame containing the features to analyze. Rows represent samples
and columns represent features.}

\item{col_samples}{A string specifying the column name in \code{df_samples} to use for sample filtering.}

\item{col_features}{A string specifying the column name in \code{df_features} to use for feature filtering.}

\item{df_samples}{A data frame containing sample data for filtering.}

\item{df_features}{A data frame containing feature data for filtering.}
}
\value{
A list containing the filtered data frame, the sample outlier plot,
and a vector of IDs of excluded samples.
}
\description{
This function performs Principal Component Analysis (PCA) on the provided
data frame and identifies outliers based on the Local Outlier Factor (LOF).
It generates plots for sample outliers and excludes identified outliers from
the original data frame.
}
\examples{
\dontrun{
  result <- outlier_pca_lof(df, col_samples = "sample_col",
                              col_features = "feature_col",
                              df_samples = df_samples,
                              df_features = df_features)
  print(result$filtered_df)
  print(result$plot_samples)
  print(result$excluded_samples)
}
}
