% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{process_data_data_check}
\alias{process_data_data_check}
\title{\code{process_data()} function: check required columns}
\usage{
process_data_data_check(
  data,
  data_meta_features,
  data_meta_samples,
  col_samples,
  col_features = NULL,
  exclusion_extreme_feature = FALSE,
  exclusion_extreme_sample = FALSE,
  col_missing_feature = NULL,
  missing_pct_feature = NULL,
  col_missing_sample = NULL,
  missing_pct_sample = NULL
)
}
\arguments{
\item{data}{Data frame containing sample data, which must include \code{col_samples}.}

\item{data_meta_features}{Data frame containing feature metadata, which may need to include
\code{col_features}, \code{col_missing_feature}, and \code{missing_pct_feature} depending on parameters.}

\item{data_meta_samples}{Data frame containing sample metadata, which may need to include
\code{col_missing_sample} and \code{missing_pct_sample} depending on parameters.}

\item{col_samples}{Character vector specifying the column name(s) in \code{data} representing sample IDs.}

\item{col_features}{Character string specifying the column name in \code{data_meta_features} representing
feature IDs (optional, only checked if provided).}

\item{exclusion_extreme_feature}{Logical; if \code{TRUE}, both \code{col_missing_feature} and \code{missing_pct_feature}
must be provided and exist in \code{data_meta_features}.}

\item{exclusion_extreme_sample}{Logical; if \code{TRUE}, both \code{col_missing_sample} and \code{missing_pct_sample}
must be provided and exist in \code{data_meta_samples}.}

\item{col_missing_feature}{Character string specifying the column name in \code{data_meta_features} for
missing data in features (required if \code{exclusion_extreme_feature} is \code{TRUE}).}

\item{missing_pct_feature}{Numeric value indicating the acceptable percentage of missing data for features
(required if \code{exclusion_extreme_feature} is \code{TRUE}).}

\item{col_missing_sample}{Character string specifying the column name in \code{data_meta_samples} for
missing data in samples (required if \code{exclusion_extreme_sample} is \code{TRUE}).}

\item{missing_pct_sample}{Numeric value indicating the acceptable percentage of missing data for samples
(required if \code{exclusion_extreme_sample} is \code{TRUE}).}
}
\value{
Logical; \code{TRUE} if all required columns are present, otherwise an error is raised.
}
\description{
This function checks for the presence of required columns
It verifies that necessary columns are present based on specified exclusion flags.
}
