% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{process_data_exclude_features}
\alias{process_data_exclude_features}
\title{Exclude features with X misingness}
\usage{
process_data_exclude_features(
  df,
  df_features,
  missing_pct_feature = NULL,
  col_missing_feature,
  col_features
)
}
\arguments{
\item{df}{A data frame containing the main data with features to be filtered.}

\item{df_features}{A data frame containing metadata for features, including missingness percentages.}

\item{missing_pct_feature}{Numeric, the threshold (between 0 and 1) for feature missing data exclusion (e.g., 0.2 for 20\%).}

\item{col_missing_feature}{Character, the column name in \code{df_features} indicating the percentage of missing data per feature.}

\item{col_features}{Character, the column name in \code{df_features} with feature IDs corresponding to columns in \code{df}.}
}
\value{
A list containing:
\item{df}{The filtered data frame with excluded features removed.}
\item{df_features}{The filtered feature metadata with excluded features removed.}
}
\description{
This function excludes features from a dataset (\code{df}) based on a specified
threshold of missing data percentage in a meta data file.
}
