% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{process_data_exclude_samples}
\alias{process_data_exclude_samples}
\title{Exclude samples with X missingness}
\usage{
process_data_exclude_samples(
  df,
  df_samples,
  missing_pct_sample = NULL,
  col_missing_sample,
  col_samples
)
}
\arguments{
\item{df}{A data frame containing the main data with samples to be filtered.}

\item{df_samples}{A data frame containing metadata for samples, including missingness percentages.}

\item{missing_pct_sample}{Numeric, the threshold (between 0 and 1) for sample missing data exclusion (e.g., 0.2 for 20\%).}

\item{col_missing_sample}{Character, the column name in \code{df_samples} indicating the percentage of missing data per sample.}

\item{col_samples}{Character, the column name in \code{df_samples} with sample IDs corresponding to row names in \code{df}.}
}
\value{
A list containing:
\item{df}{The filtered data frame with excluded samples removed.}
\item{df_samples}{The filtered sample metadata with excluded samples removed.}
\item{excluded_samples}{A vector of the names of samples that were excluded.}
}
\description{
This function excludes samples from a dataset (\code{df}) based on a specified
threshold of missing data percentage in a meta data file.
}
