% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{process_data_generate_labels}
\alias{process_data_generate_labels}
\title{\code{process_data()} function: generate labels}
\usage{
process_data_generate_labels(
  exclusion_extreme_feature,
  missing_pct_feature = NULL,
  exclusion_extreme_sample,
  missing_pct_sample = NULL,
  imputation,
  imputation_method = NULL,
  transformation,
  transformation_method = NULL,
  outlier,
  plate_correction,
  centre_scale
)
}
\arguments{
\item{exclusion_extreme_feature}{Logical, whether extreme feature exclusion is applied.}

\item{missing_pct_feature}{Numeric, the percentage threshold for excluding features based on missing data (only used if \code{exclusion_extreme_feature} is TRUE).}

\item{exclusion_extreme_sample}{Logical, whether extreme sample exclusion is applied.}

\item{missing_pct_sample}{Numeric, the percentage threshold for excluding samples based on missing data (only used if \code{exclusion_extreme_sample} is TRUE).}

\item{imputation}{Logical, whether imputation is applied.}

\item{imputation_method}{Character, the method used for imputation (only used if \code{imputation} is TRUE).}

\item{transformation}{Logical, whether data transformation is applied.}

\item{transformation_method}{Character, the method used for transformation (only used if \code{transformation} is TRUE).}

\item{outlier}{Logical, whether outlier handling is applied.}

\item{plate_correction}{Logical, whether plate correction is applied.}

\item{centre_scale}{Logical, whether centering and scaling are applied.}
}
\value{
A list of character labels indicating the settings for each parameter.
}
\description{
This function creates labels for saving based on several data processing options
(e.g., exclusion, imputation, transformation, outlier handling, etc.).
It returns a list of character labels that indicate the specified settings,
which can be used for file naming or tracking parameter choices.
}
