% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{process_feature_data}
\alias{process_feature_data}
\title{Process and Normalize feature data}
\usage{
process_feature_data(
  data,
  data_meta_features,
  data_meta_samples,
  col_samples,
  col_features,
  path_out,
  path_outliers,
  exclusion_extreme = FALSE,
  col_missing = NULL,
  missing_pct = NULL,
  imputation = FALSE,
  imputation_method = NULL,
  col_LOD = NULL,
  transformation = FALSE,
  transformation_method = NULL,
  outlier = FALSE,
  plate_correction = FALSE,
  cols_listRandom = NULL,
  cols_listFixedToKeep = NULL,
  cols_listFixedToRemove = NULL,
  col_HeteroSked = NULL,
  case_control = FALSE,
  col_case_control = NULL
)
}
\arguments{
\item{data}{A data frame with col1 sample ID and remaining columns of features with feature IDs as colnames}

\item{data_meta_features}{A data frame containing metadata for the features}

\item{data_meta_samples}{A data frame containing metadata for the samples}

\item{col_samples}{A string specifying the column name in \code{data} that contains sample IDs. For example, \code{"Idepic_Bio"}.}

\item{col_features}{A string specifying the column name in \code{data} that contains feature IDs (should be the same values as the colnames of \code{data}. For example, \code{"UNIPROT"}.}

\item{path_out}{A string specifying the output directory where the processed data will be saved}

\item{path_outliers}{A string specifying the output directory where outlier information will be saved}

\item{exclusion_extreme}{A logical flag to indicate if features with extreme missingness should be excluded. Default is \code{FALSE}.}

\item{col_missing}{A string specifying the column name in \code{data_meta_features} that contains the percentage of missing values for each feature. For example, \code{"missing_pct"}.}

\item{missing_pct}{A numeric value specifying the threshold percentage for missingness above which features will be excluded. For example, \code{0.9}.}

\item{imputation}{A logical flag to indicate if imputation should be performed. Default is \code{FALSE}.}

\item{imputation_method}{A string specifying the method to use for imputation. Options are \code{"LOD"}, \code{"1/5th"}, \code{"KNN"}, \code{"ppca"}, \code{"median"}, \code{"mean"}, \code{"rf"}, and \code{"left-censored"}.}

\item{col_LOD}{A string specifying the column name in \code{data_meta_features} that contains the limit of detection (LOD) values, required if \code{imputation_method} is \code{"LOD"}.}

\item{transformation}{A logical flag to indicate if transformation should be performed. Default is \code{FALSE}.}

\item{transformation_method}{A string specifying the method to use for transformation. Options are \code{"InvRank"}, \code{"Log10"}, \code{"Log10Capped"}, and \code{"Log10ExclExtremes"}.}

\item{outlier}{A logical flag to indicate if outlier exclusion should be performed across features and samples. Default is \code{FALSE}.}

\item{plate_correction}{A logical flag to indicate if plate correction should be performed. Default is \code{FALSE}.}

\item{cols_listRandom}{A string or vector specifying columns to be treated as random effects in plate correction. For example, \code{"batch_plate"}.}

\item{cols_listFixedToKeep}{A vector specifying columns to be treated as fixed effects in plate correction and kept in the model. For example, \code{c("Center", "Country")}.}

\item{cols_listFixedToRemove}{A vector specifying columns to be treated as fixed effects in plate correction and removed from the model.}

\item{col_HeteroSked}{A string specifying the column to be used for heteroskedasticity correction.}

\item{case_control}{A logical flag to indicate if the data are case-control and matched samples should be handled accordingly. Default is \code{FALSE}.}

\item{col_case_control}{A string specifying the column name in \code{data_meta_samples} that contains case-control matching information. For example, \code{"Match_Caseset"}.}
}
\value{
The processed and normalized data is saved to the specified output directory.
}
\description{
This function processes a given dataset with specified metadata features and samples. It supports exclusion of features with extreme missingness, various imputation methods, transformation, plate correction, and case-control data handling.
}
\details{
This function performs several data processing steps:
\itemize{
\item Excludes features with extreme missingness based on a specified threshold.
\item Imputes missing values using various methods.
\item Transforms the data using specified methods.
\item Corrects for plate effects using specified random and fixed effects.
\item Handles case-control data to ensure matched samples are treated appropriately.
}
}
