% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{proxy_search_split}
\alias{proxy_search_split}
\title{finds proxy SNPs from an internal reference
split the SNP search into 10k blocks}
\usage{
proxy_search_split(
  data_exposure,
  data_outcome,
  data_outcome_path,
  data_reference,
  data_reference_path,
  tag_r2 = 0.8,
  tag_kb = 5000,
  tag_nsnp = 5000,
  outcome_sep,
  outcome_phenotype,
  outcome_SNP,
  outcome_BETA,
  outcome_SE,
  outcome_P,
  outcome_EA,
  outcome_OA,
  outcome_EAF,
  outcome_N,
  outcome_ID,
  outcome_CHR,
  outcome_POS
)
}
\arguments{
\item{data_exposure}{Data frame containing exposure SNP data.}

\item{data_outcome}{Data frame containing outcome SNP data.}

\item{data_outcome_path}{Path to the outcome data file.}

\item{data_reference}{Data frame containing reference SNP data.}

\item{data_reference_path}{Path to the reference data file.}

\item{tag_r2}{Threshold for LD (default: 0.8).}

\item{tag_kb}{Threshold for distance in kb (default: 5000).}

\item{tag_nsnp}{Threshold for number of SNPs (default: 5000).}

\item{outcome_sep}{Delimiter used in the outcome data file.}

\item{outcome_phenotype}{Column name for the outcome phenotype.}

\item{outcome_SNP}{Column name for the SNP in the outcome data.}

\item{outcome_BETA}{Column name for the beta value in the outcome data.}

\item{outcome_SE}{Column name for the standard error in the outcome data.}

\item{outcome_P}{Column name for the p-value in the outcome data.}

\item{outcome_EA}{Column name for the effect allele in the outcome data.}

\item{outcome_OA}{Column name for the other allele in the outcome data.}

\item{outcome_EAF}{Column name for the effect allele frequency in the outcome data.}

\item{outcome_N}{Column name for the sample size in the outcome data.}

\item{outcome_ID}{Column name for the ID in the outcome data.}

\item{outcome_CHR}{Column name for the chromosome in the outcome data.}

\item{outcome_POS}{Column name for the position in the outcome data.}
}
\value{
Data frame with proxy SNPs added.
}
\description{
finds proxy SNPs from an internal reference
split the SNP search into 10k blocks
}
