% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{proxy_search_UKB}
\alias{proxy_search_UKB}
\title{finds proxy SNPs from an internal reference (UKB format)}
\usage{
proxy_search_UKB(
  data_exposure,
  data_outcome,
  data_outcome_path,
  data_reference,
  data_reference_path,
  tag_r2 = 0.8,
  tag_kb = 5000,
  tag_nsnp = 5000,
  outcome_sep
)
}
\arguments{
\item{data_exposure}{Data frame containing exposure SNP data.}

\item{data_outcome}{Data frame containing outcome SNP data.}

\item{data_outcome_path}{Path to the outcome data file.}

\item{data_reference}{Data frame containing reference SNP data.}

\item{data_reference_path}{Path to the reference data file.}

\item{tag_r2}{Threshold for LD (default: 0.8).}

\item{tag_kb}{Threshold for distance in kb (default: 5000).}

\item{tag_nsnp}{Threshold for number of SNPs (default: 5000).}

\item{outcome_sep}{Delimiter used in the outcome data file.}
}
\value{
Data frame with proxy SNPs added.
}
\description{
finds proxy SNPs from an internal reference (UKB format)
}
