% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{qqplot}
\usage{
qqplot(
  df,
  ci = 0.95,
  point_size = 2,
  point_colour = "black",
  point_alpha = 0.8
)
}
\arguments{
\item{df}{A data frame AND p-value column i.e., data$P}

\item{ci}{A number between 0 and 1 (defaults to 0.95) indicating the type of confidence interval to be drawn. Use NA to remove.}

\item{point_size}{the size of the points - defaults to 2}

\item{point_colour}{the colour of the points - defaults to black}

\item{point_alpha}{the alpha/transparency of the points - defaults to 0.8}
}
\description{
qqplot
}
