% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR-edits.R
\name{random_string}
\alias{random_string}
\title{Generate random strings}
\usage{
random_string(n = 1, len = 6)
}
\arguments{
\item{n}{Number of random strings to generate.}

\item{len}{Length of each random string.}
}
\value{
A vector of randomly generated strings.
}
\description{
This function generates random strings of specified length.
}
