% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_file}
\alias{read_file}
\title{Read File Based on Extension}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the file.}
}
\value{
A data frame or data table containing the data read from the file.
}
\description{
This function reads a file based on its extension and returns the data in a suitable format.
It supports reading text files, CSV files, Stata files, Excel files, and SAS files.
}
\details{
The function uses different packages to read different file types:
\itemize{
\item \code{data.table::fread} for reading text and CSV files.
\item \code{readstata13::read.dta13} for reading Stata files.
\item \code{readxl::read_xlsx} for reading Excel files.
\item \code{haven::read_sas} for reading SAS files.
}
}
