% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{recode_indels_21}
\alias{recode_indels_21}
\title{Recodes indels for 2-1 allele configuration.}
\usage{
recode_indels_21(A1, A2, B1)
}
\arguments{
\item{A1}{First allele for exposure.}

\item{A2}{Second allele for exposure.}

\item{B1}{First allele for outcome.}
}
\value{
A data frame with recoded alleles.
}
\description{
This function recodes insertions and deletions (indels) for a 2-1 allele configuration.
}
