% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{remove_duplicate_SNP}
\alias{remove_duplicate_SNP}
\title{Remove duplicate rows based on SNP column}
\usage{
remove_duplicate_SNP(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A modified data frame with duplicate rows removed.
\dontrun{
for (i in seq_along(list_data_cis_trans)) {
df <- list_data_cis_trans[\link{i}]
df <- remove_duplicate_SNP(df)
df <- remove_nearby_positions(df)
list_data_cis_trans[\link{i}] <- df
}
}
\description{
This function removes duplicate rows based on the SNP column,
keeping only the rows with "cis" in the "cis_trans" column.
It iterates through each row with "cis" in the "cis_trans" column,
checks if there are duplicate rows with the same value in the SNP column,
and removes the duplicates, keeping the row with "cis" in the "cis_trans" column.
}
