% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{remove_nearby_positions}
\alias{remove_nearby_positions}
\title{Remove rows with nearby positions}
\usage{
remove_nearby_positions(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A modified data frame with rows removed based on nearby positions.
\dontrun{
for (i in seq_along(list_data_cis_trans)) {
df <- list_data_cis_trans[\link{i}]
df <- remove_duplicate_SNP(df)
df <- remove_nearby_positions(df)
list_data_cis_trans[\link{i}] <- df
}
}
\description{
This function removes rows with nearby positions, keeping only the rows with "cis"
in the "cis_trans" column. It iterates through each row with "cis" in the "cis_trans" column,
checks if there are other rows with the same value in the "chr.exposure" column,
and if the positions are within ±500,000. If such rows are found, it removes the rows
without "cis" in the "cis_trans" column.
}
