% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{susieR_cs_table}
\alias{susieR_cs_table}
\title{Create a Table of Credible Sets from SusieR Model}
\usage{
susieR_cs_table(model)
}
\arguments{
\item{model}{A \code{susieR} model object, which includes the following components:
\itemize{
\item \code{sets} A list containing credible sets, where each set is represented by indices of SNPs.
\item \code{X_column_scale_factors} A vector of SNP identifiers (e.g., SNP names).
\item \code{pip} A named vector of Posterior Inclusion Probability (PIP) values, indexed by SNP names.
}}
}
\value{
A tibble (data frame) containing the following columns:
\itemize{
\item \code{SNP} The SNP identifiers for each SNP in the credible set.
\item \code{PIP} The Posterior Inclusion Probability for each SNP.
\item \code{cs_snps} A string listing the other SNPs in the same credible set (NA if only one SNP).
\item \code{test} The label "susie", indicating the test used for generating the table.
}
}
\description{
This function processes a \code{susieR} model object to extract the SNPs and their corresponding Posterior Inclusion Probabilities (PIP) for each credible set. It also includes a list of other SNPs in the same credible set.
}
\examples{
\dontrun{
  result <- susieR_cs_table(susie_model)
  print(result)
}
}
