% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-feature-data.R
\name{transform_data}
\alias{transform_data}
\title{Apply Data Transformation to Data Frame}
\usage{
transform_data(df, transformation_method)
}
\arguments{
\item{df}{A data frame containing numeric data for transformation.}

\item{transformation_method}{A string specifying the transformation method to apply.
Supported values are "InvRank", "Log10", "Log10Capped", and "Log10ExclExtremes".}
}
\value{
A list containing:
\item{df}{The transformed data frame with the applied method.}
\item{transformation_method}{The label of the applied transformation method.}
\item{centre_scale}{A label indicating whether the data was centered and scaled (TRUE or NA).}
If conditions are not met for the specified transformation method, the original data frame will be returned in the list.
}
\description{
This function applies various data transformations to a numeric data frame based on a specified transformation method.
Supported methods include "InvRank", "Log10", "Log10Capped", and "Log10ExclExtremes".
}
