% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_plot.R
\name{blank_plot}
\alias{blank_plot}
\title{Create a blank ggplot2 plot}
\usage{
blank_plot(message)
}
\arguments{
\item{message}{Optional. A character string to be displayed as a message within the plot. Defaults to NULL (no message).}
}
\value{
A ggplot2 plot object
}
\description{
This function creates a blank ggplot2 plot with no data or aesthetics. It can be used as a placeholder
or to customize axes and labels without displaying any data points.
}
