% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{format_data}
\alias{format_data}
\title{Format Data for Mendelian Randomization Analysis}
\usage{
format_data(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{A data frame containing the input data.}

\item{type}{A character string specifying the name of the phenotype column. Default is "exposure".}

\item{snps}{A vector of SNP IDs to subset the data. Default is NULL.}

\item{header}{Logical indicating whether the data has a header row. Default is TRUE.}

\item{phenotype_col}{A character string specifying the name of the phenotype column in the input data. Default is "Phenotype".}

\item{snp_col}{A character string specifying the name of the SNP column in the input data. Default is "SNP".}

\item{beta_col}{A character string specifying the name of the beta coefficient column in the input data. Default is "beta".}

\item{se_col}{A character string specifying the name of the standard error column in the input data. Default is "se".}

\item{eaf_col}{A character string specifying the name of the effect allele frequency column in the input data. Default is "eaf".}

\item{effect_allele_col}{A character string specifying the name of the effect allele column in the input data. Default is "effect_allele".}

\item{other_allele_col}{A character string specifying the name of the other allele column in the input data. Default is "other_allele".}

\item{pval_col}{A character string specifying the name of the p-value column in the input data. Default is "pval".}

\item{units_col}{A character string specifying the name of the units column in the input data. Default is "units".}

\item{ncase_col}{A character string specifying the name of the number of cases column in the input data. Default is "ncase".}

\item{ncontrol_col}{A character string specifying the name of the number of controls column in the input data. Default is "ncontrol".}

\item{samplesize_col}{A character string specifying the name of the sample size column in the input data. Default is "samplesize".}

\item{gene_col}{A character string specifying the name of the gene column in the input data. Default is "gene".}

\item{id_col}{A character string specifying the name of the ID column in the input data. Default is "id".}

\item{min_pval}{A numeric value specifying the minimum p-value threshold. Default is 1e-200.}

\item{z_col}{A character string specifying the name of the Z-score column in the input data. Default is "z".}

\item{info_col}{A character string specifying the name of the INFO column in the input data. Default is "info".}

\item{chr_col}{A character string specifying the name of the chromosome column in the input data. Default is "chr".}

\item{pos_col}{A character string specifying the name of the position column in the input data. Default is "pos".}

\item{log_pval}{Logical indicating whether to transform p-values to -log10 scale. Default is FALSE.}
}
\value{
A formatted data frame ready for Mendelian Randomization analysis.
}
\description{
This function formats input data for Mendelian Randomization (MR) analysis.
It extracts columns based on specified column names, cleans and subsets the data,
handles phenotype column, transforms p-values if required, removes duplicated SNPs,
and performs additional checks and transformations required for MR analysis.
}
