% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy-search.R
\name{proxy_search}
\alias{proxy_search}
\title{proxy_search: search for proxies when using read_outcome_data()}
\usage{
proxy_search(
  data_exposure,
  data_outcome,
  data_outcome_path,
  data_reference_path
)
}
\arguments{
\item{data_exposure}{exposure data frame}

\item{data_outcome}{outcome data frame}

\item{data_outcome_path}{file path used for read_outcome_data()}

\item{data_reference_path}{file path for your downloaded reference panel}
}
\value{
a data frame
}
\description{
proxy_search: search for proxies when using read_outcome_data()
}
