% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{remove_duplicate_SNP}
\alias{remove_duplicate_SNP}
\title{Remove duplicate rows based on SNP column}
\usage{
remove_duplicate_SNP(df)
}
\arguments{
\item{df}{A data frame containing SNP and cis_trans columns.}
}
\value{
A modified data frame with duplicate rows removed based on SNP, prioritizing "cis" rows.
}
\description{
This function removes duplicate rows based on the "SNP" column, prioritizing rows with "cis" in the "cis_trans" column.
For a given SNP, it keeps only one row, prioritizing the row with "cis" in "cis_trans" and removing others.
}
\details{
Example usage:
\code{#' for (i in seq_along(list_data_cis_trans)) {
  df <- list_data_cis_trans[[i]]
  df <- remove_duplicate_SNP(df)
  df <- remove_nearby_positions(df)
  list_data_cis_trans[[i]] <- df
}
}
}
