% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR_functions.R
\name{remove_nearby_positions}
\alias{remove_nearby_positions}
\title{Remove rows with nearby positions (excluding "cis" and "cis-trans_cis")}
\usage{
remove_nearby_positions(df)
}
\arguments{
\item{df}{A data frame containing chr.exposure, pos.exposure, and cis_trans columns.}
}
\value{
A modified data frame with rows removed based on nearby positions (excluding "cis" and "cis-trans_cis").
}
\description{
This function removes rows with positions within ±1mb of a "cis" or "cis-trans_cis" row,
excluding the "cis" or "cis-trans_cis" row itself. It iterates through rows with "cis" or "cis-trans_cis" in "cis_trans",
checks for other rows with the same chromosome but positions within the threshold.
If such rows exist and their "cis_trans" is not "cis" or "cis-trans_cis", it removes those rows.
}
\details{
Example usage:
\code{
for (i in seq_along(list_data_cis_trans)) {
  df <- list_data_cis_trans[[i]]
  df <- remove_duplicate_SNP(df)
  df <- remove_nearby_positions(df)
  list_data_cis_trans[[i]] <- df
}
}
}
